requirements-server.txt
# LoRa-hX 수신 서버 의존성
pyserial>=3.5
pyyaml>=6.0
peewee>=3.17
fastapi>=0.104
uvicorn>=0.24
websockets>=12.0
